Changed timer variables from unsigned to signed, implemented LED funcs
Signed time is so that even if the player uses his whole time the time
continues to run so at the end of the game it is visible how slow was
that player.
RGB LED lights up in the color of the current player, BLACK is
represented with purple-white color.